+++ script/check_release_configuration.sh
@@ -138,6 +138,7 @@ for resource in \
   "$ROOT_DIR/SurgeRelay/WebResources/web-markup.js" \
   "$ROOT_DIR/SurgeRelay/WebResources/web-api.js" \
   "$ROOT_DIR/SurgeRelay/WebResources/web-state.js" \
+  "$ROOT_DIR/SurgeRelay/WebResources/web-editor.js" \
   "$ROOT_DIR/SurgeRelay/WebResources/app.js"
 do
   require_file "$resource"